<fix> : Build warning solved - warning C4244: 'argument' : conversion from 'std::streamsize' to 'unsigned int', possible loss of data

+++ SFXTemplete/SFXTemplete.cpp
@@ -85,7 +85,7 @@ BOOL CSFXTempleteApp::ExtractEmbeddedZip(LPCTSTR outputPath)
 		return FALSE;
 	}
 
-	std::vector<char> zipData(zipSize);
+	std::vector<char> zipData((int)zipSize);
 	exeFile.read(zipData.data(), zipSize);
 	exeFile.close();
